// frontend/src/models/portfolioPositionScheme.ts
/**
 * Generated by orval v6.10.3 🍺
 * Do not edit manually.
 * Invest API
 * OpenAPI spec version: 0.1.0
 */
import type { InstrumentScheme } from './instrumentScheme';

export interface PortfolioPositionScheme {
  instrument: InstrumentScheme;
  quantity: number;
  blocked: number;
  average_price: number;
  expected_yield: number;
  current_price: number;
  var_margin: number;
  current_nkd: number;
}